R backendCommits

+++ src/main/java/com/group/security/Exception/EventNotFoundException.java
@@ -1,7 +1,7 @@
 package com.group.security.Exception;
 
 public class EventNotFoundException extends RuntimeException{
-    public EventNotFoundException(Long id){
+    public EventNotFoundException(Integer id){
         super("Event not found id"+id);
     }
 }
